basic crypto working (single client, fixed keys)
basic functionality is working.
encryption is done in Echo::send and decryption in worker::handleEchoData (virtual for client & server)
for the time keys are fixed and hardcoded.

key should be derived from passphrase (keystreching) and nonce is random generated on challenge request
and then increased (if 8byte char or 24 byte char is used is still to be decided)

Utility's rand function should use real rand (/dev/random) has to be adjusted for classes (mainly Auth)

nonce should be sent on challenge request (isConnectionRequest variable in echo)