increase nonce with each message
nonce must be only used once.
We already send a random nonce on beginning of authentication.
on every packet received increase nonce to ensure it is used only once.

salsa20 cipher has a 8byte nonce which is enough when each nonce ist not chosen randomly,
so every time we receive or sent a packet we increase nonce.

TODO: to keep track of lost packets or out of order packets we keep a reference between
nonce and seq. therefore change echo sequence is enforced.