create encryption key from password
password is created from given passphrase now.

TODO: a simple keystrech algorithm is used. check for better options.
example: scrypt .. but we do not want to add a lot of dependencies.
evaluate keystrength possibilities. code should be easily changeable.
Auth::Auth() needs to be changed when better option is found