compensate for packet loss in nonce counting
when a packet is lost the nonce counting would be out of order.
to compensate for this we store the last received seq.
when a packet is received we subtract stored seq from act seq and use
this value as nonce increment. (when no packet is lost this is 1)

also: use random as ICMP header ID and increment ICMP header SEQ on every echo.
(linux behaviour)