handle multiuser & multiuser nat
echo ID is used as client identifier. Since seq and ID is not encrypted
echo ID is chosen random on beginning and never changed. This is Linux behavior
which uses PID as ID and increases SEQ on every echo (as we do)

with this it should be possible to have multiple clients even behind nat